Issues #82 test: GifticonParser 수정 및 TestCode 추가
